Split the OpenAL sound code into common and OpenAL-specific parts to make it easier to add support for other audio APIs.
