+++ common/CMakeLists.txt
@@ -134,7 +134,7 @@ if(DSOUND)
     list(APPEND COMMONV_SRC soundio.cpp vqaaudio_dsound.cpp)
     list(APPEND VANILLA_LIBS dsound)
 elseif(OPENAL)
-    list(APPEND COMMONV_SRC soundio_openal.cpp vqaaudio_openal.cpp)
+    list(APPEND COMMONV_SRC soundio_common.cpp soundio_openal.cpp vqaaudio_openal.cpp)
 else()
     list(APPEND COMMONV_SRC soundio_null.cpp vqaaudio_null.cpp)
 endif()
